Add timestamp just for keeping accurate timestamps; cleanup

+++ src/dildonica_midi.hpp
@@ -2,5 +2,5 @@
 struct DildonicaData {
 	uint32_t timestamp;
 	uint32_t rawSample;
-    uint32_t zone;
+    uint8_t zone;
 };
